% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ArgCheck.R
\name{addError}
\alias{addError}
\title{Argcheck}
\usage{
addError(msg, argcheck)
}
\arguments{
\item{msg}{Text to show.}

\item{argcheck}{Environment created by \code{\link{newArgCheck}}.}
}
\value{
Add an error in the environment \code{argcheck}.
}
\description{
Add an error in the environment created by \code{\link{newArgCheck}}.
}
\keyword{internal}
