% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ArgCheck.R
\name{addWarning}
\alias{addWarning}
\title{Argcheck}
\usage{
addWarning(msg, argcheck)
}
\arguments{
\item{msg}{Text to show.}

\item{argcheck}{Environment created by \code{\link{newArgCheck}}.}
}
\value{
Add a warning in the environment \code{argcheck}.
}
\description{
Add a warning in the environment created by \code{\link{newArgCheck}}.
}
\keyword{internal}
