% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cst_deliv_duration.R
\encoding{UTF-8}
\name{cst_deliv_duration.verif_args}
\alias{cst_deliv_duration.verif_args}
\title{Verification}
\usage{
cst_deliv_duration.verif_args(
  Rx_deliv,
  Rx_drug_code,
  Rx_deliv_dur,
  Cst_deliv_dur,
  Cst_drug_code,
  Cst_duration
)
}
\description{
Arguments verification for \code{\link{cst_deliv_duration}}.
}
\keyword{internal}
