% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cst_tx_duration.R
\name{cst_tx_duration}
\alias{cst_tx_duration}
\title{Constant treatment duration drugs}
\usage{
cst_tx_duration(
  Rx_deliv,
  Rx_drug_code,
  Rx_duration,
  cst_tx_dur,
  cst_drug_code,
  cst_duration
)
}
\arguments{
\item{Rx_deliv}{Name of the table listing all prescription drugs delivered.}

\item{Rx_drug_code}{Column name of \code{Rx_deliv} that contains the drug unique identifier.}

\item{Rx_duration}{Column name of the constant treatment duration in the \code{Rx_deliv} table.}

\item{cst_tx_dur}{Name of the table that contains the constant treatment durations that will overwrite that in the \code{Rx_deliv} table for the specified drug codes.}

\item{cst_drug_code}{Column name of \code{cst_tx_dur} that contains the drug unique identifier (same format as \code{Rx_drug_code}).}

\item{cst_duration}{Column name of the constant treatment duration in the \code{cst_tx_dur} table (same format as \code{Rx_duration}).}
}
\value{
\code{data.table} of the same structure than \code{Rx_deliv} listing all drugs in which a constant treatment duration replaces the original treatment duration.
}
\description{
Overwrites the treatment duration with constant durations for each drug code included in a user-provided table.
}
\examples{
Rx_dt <- data.frame(id = c(rep(1, 3), rep(2, 2)),
                    code = c("A", "B", "C", "B", "D"),
                    duration = c(rep(15, 3), 15, 90))
cst_dt <- data.frame(codes = c("A", "C", "D"),
                     dur = c(50, 100, 45))
Rx_cst <- cst_tx_duration(Rx_deliv = Rx_dt,
                          Rx_drug_code = "code", Rx_duration = "duration",
                          cst_tx_dur = cst_dt,
                          cst_drug_code = "codes", cst_duration = "dur")
}
