% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_process.R
\encoding{UTF-8}
\name{data_process.1_core}
\alias{data_process.1_core}
\title{Data Process}
\usage{
data_process.1_core(
  Rx_deliv,
  Rx_id,
  Rx_drug_code,
  Rx_drug_deliv,
  Rx_deliv_dur,
  Cohort = NULL,
  Cohort_id = NULL,
  Hosp_stays = NULL,
  Hosp_id = NULL,
  Hosp_admis = NULL,
  Hosp_discharge = NULL,
  study_start = NULL,
  study_end = NULL,
  grace_fctr = 0.5,
  grace_cst = 0,
  max_reserve = NULL
)
}
\description{
\code{\link{data_process}} but with only 1 core. To use in the multicores process.
}
\keyword{internal}
