% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_process.R
\encoding{UTF-8}
\name{data_process.verif_args}
\alias{data_process.verif_args}
\title{Verification}
\usage{
data_process.verif_args(
  Rx_deliv,
  Rx_id,
  Rx_drug_code,
  Rx_drug_deliv,
  Rx_deliv_dur,
  Cohort,
  Cohort_id,
  Hosp_stays,
  Hosp_id,
  Hosp_admis,
  Hosp_discharge,
  study_start,
  study_end,
  grace_fctr,
  grace_cst,
  max_reserve,
  verif_cols,
  cores
)
}
\description{
Arguments verification for \code{\link{data_process}}.
}
\keyword{internal}
