% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-functions.R
\encoding{UTF-8}
\name{rmNA}
\alias{rmNA}
\title{Utils}
\usage{
rmNA(x)
}
\arguments{
\item{x}{Vector.}
}
\value{
\code{x} without \code{NA}s or \code{NaN}s.
}
\description{
Delete \code{NA}s and \code{NaN}s.
}
\examples{
x <- c(1:10, NA, NaN)
rmNA(x)
}
\keyword{internal}
