% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-data.R
\docType{data}
\name{Rx_processed}
\alias{Rx_processed}
\title{Table: Processed "unprocessed table"}
\format{
A \code{data.table} with 6792 obs and 4 variables:
\describe{
\item{id}{Individual unique identifier.}
\item{code}{Drug unique identifier.}
\item{tx_start}{The date of initiation of the reconstructed continued treatment (format as date).}
\item{tx_end}{The date of the last day of the reconstructed continued treatment (format as date).}
}
}
\usage{
Rx_processed
}
\description{
Table required for the calculation of the polypharmacy indicators. This table is created by using \code{data_process()} function on \code{Rx_unprocessed} data.
}
\keyword{datasets}
