% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-data.R
\docType{data}
\encoding{UTF-8}
\name{Rx_unprocessed}
\alias{Rx_unprocessed}
\title{Table: Prescription drugs deliveries}
\format{
A \code{data.table} with 17060 obs and 4 variables:
\describe{
\item{id}{Individual unique identifier}
\item{code}{Drug unique identifier}
\item{start}{Date of the drug delivery}
\item{duration}{Duration of the delivery}
}
}
\usage{
Rx_unprocessed
}
\description{
Table listing all prescription drugs deliveries.
}
\keyword{datasets}
