% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-data.R
\docType{data}
\encoding{UTF-8}
\name{sample_Rx_processed}
\alias{sample_Rx_processed}
\title{Table: Processed \emph{unprocessed table}}
\format{
A \code{data.table} with 6792 obs and 4 variables:
\describe{
\item{id}{Individual unique identifier.}
\item{code}{Medication unique identifier.}
\item{tx_start}{The date of initiation of the reconstructed continued treatment (format as date).}
\item{tx_end}{The end date of the reconstructed continued treatment (format as date).}
}
}
\usage{
sample_Rx_processed
}
\description{
This table is provided to users of this package for training purposes. It is created by using \code{\link{data_process}} function on \code{\link{sample_Rx_unprocessed}} data.
}
\keyword{datasets}
