% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-data.R
\docType{data}
\encoding{UTF-8}
\name{sample_Rx_unprocessed}
\alias{sample_Rx_unprocessed}
\title{Table: Prescription drugs deliveries}
\format{
A \code{data.table} with 17060 obs and 4 variables:
\describe{
\item{id}{Individual unique identifier.}
\item{code}{Medication unique identifier.}
\item{start}{Date of the medication delivery.}
\item{duration}{Treatment duration of the delivery.}
}
}
\usage{
sample_Rx_unprocessed
}
\description{
A sample table of prescription drugs deliveries provided to users of this package for training purposes. It contains the raw information that leads to \code{\link{sample_Rx_processed}} when processes by the \code{\link{data_process}} function.
}
\keyword{datasets}
