% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cst_deliv_duration.R
\name{cst_deliv_duration}
\alias{cst_deliv_duration}
\title{Constant delivery duration drugs}
\usage{
cst_deliv_duration(
  Rx_deliv,
  Rx_drug_code,
  Rx_deliv_dur,
  Cst_deliv_dur,
  cst_drug_code,
  cst_deliv_dur
)
}
\arguments{
\item{Rx_deliv}{Name of the table listing all prescription drugs delivered.}

\item{Rx_drug_code}{Column name of \code{Rx_deliv} that contains the drug unique identifier.}

\item{Rx_deliv_dur}{Column name of the constant treatment duration in the \code{Rx_deliv} table.}

\item{Cst_deliv_dur}{Name of the table that contains the constant delivery durations that will overwrite that in the \code{Rx_deliv} table for the specified drug codes.}

\item{cst_drug_code}{Column name of \code{Cst_deliv_dur} that contains the drug unique identifier (same format as \code{Rx_drug_code}).}

\item{cst_deliv_dur}{Column name of the constant treatment duration in the \code{Cst_deliv_dur} table (same format as \code{Rx_deliv_dur}).}
}
\value{
\code{data.table} of the same structure as \code{Rx_deliv}.
}
\description{
Overwrites the delivery durations with constant durations for each drug code listed in a user-provided table.
}
\examples{
Rx_dt <- data.frame(id = c(rep(1, 3), rep(2, 2)),
                    code = c("A", "B", "C", "B", "D"),
                    duration = c(rep(15, 3), 15, 90))
cst_dt <- data.frame(codes = c("A", "C", "D"),
                     dur = c(50, 100, 45))
Rx_cst <- cst_deliv_duration(Rx_deliv = Rx_dt,
                             Rx_drug_code = "code", Rx_deliv_dur = "duration",
                             Cst_deliv_dur = cst_dt,
                             cst_drug_code = "codes", cst_deliv_dur = "dur")
}
