% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_process.R
\name{data_process}
\alias{data_process}
\title{Create the table of the drug treatments}
\usage{
data_process(
  Rx_deliv,
  Rx_id,
  Rx_drug_code,
  Rx_drug_deliv,
  Rx_deliv_dur,
  Cohort = NULL,
  Cohort_id = NULL,
  Hosp_stays = NULL,
  Hosp_id = NULL,
  Hosp_admis = NULL,
  Hosp_discharge = NULL,
  study_start = NULL,
  study_end = NULL,
  grace_fctr = 0.5,
  grace_cst = 0,
  max_reserve = NULL
)
}
\arguments{
\item{Rx_deliv}{Name of the table listing all prescription drugs delivered including the run-in period See \emph{Details}.}

\item{Rx_id}{Column name of \code{Rx_deliv} containing individual’s unique identifier (any format).}

\item{Rx_drug_code}{Column name of \code{Rx_deliv} that contains the drug’s unique identifier (any format).}

\item{Rx_drug_deliv}{Column name of \code{Rx_deliv} that contains the dates of the drug deliveries (Date format, see \emph{Details}).}

\item{Rx_deliv_dur}{Column name of \code{Rx_deliv} that contains the duration of the delivery (integer number).}

\item{Cohort}{Name of the table providing the unique identifiers of the study cohort. Only the ids listed in both the \code{Cohort} and the \code{Rx_deliv} tables will be returned. if \code{Cohort = NULL}, all ids of the \code{Rx_deliv} table will be returned.}

\item{Cohort_id}{Column name of \code{Cohort} containing individual’s unique identifiers (same format as \code{Rx_id}). If \code{Cohort} is not \code{NULL} and \code{Cohort_id} is \code{NULL}, \code{Cohort_id} will take the same value as \code{Rx_id}.}

\item{Hosp_stays}{Name of the table listing all hospital stays. (see \emph{Details} for possible format).}

\item{Hosp_id}{Column name of \code{Hosp_stays} containing individuals’ unique identifiers (same format as \code{Rx_id}). If \code{Hosp_stays} is not \code{NULL} and \code{Hosp_id} is \code{NULL}, \code{Hosp_id} will take the same value as \code{Rx_id}.}

\item{Hosp_admis}{Column name of \code{Hosp_stays} that contains the date of admission in hospital (Date format, see \emph{Details}).}

\item{Hosp_discharge}{Column name of Hosp_stays that contains the date of discharge from hospital (Date format, see \emph{Details}).}

\item{study_start, study_end}{Defines the first and last day of the study period for which the polypharmacy indicator(s) need to be calculated. All treatment periods prior to \code{study_start} and past \code{study_end} are not transcribed into the result table (Date format, see \emph{Details}).}

\item{grace_fctr, grace_cst}{Numbers \eqn{\ge} 0. Two types of grace periods can be applied. One is proportional to the treatment duration of the latest delivery (\code{grace_fctr}) and the other is a constant number of days (\code{grace_cst}).}

\item{max_reserve}{An integer number \eqn{\ge} 0 or \code{NULL}. Longest treatment duration, in days, that can be stored from successive overlapping deliveries. When \code{max_reserve = NULL} no limit is applied. When \code{max_reserve = 0} no accumulation of extra treatment duration is accounted for.}
}
\value{
\code{data.table} with four (4) variables:
\itemize{
\item The individual unique identifier which name is defined by \code{Rx_id}.
\item The drug unique identifier which name is defined by \code{Rx_drug_code}.
\item tx_start: The date of initiation of the reconstructed continued treatment (format as date).
\item tx_end: The date of the last day of the reconstructed continued treatment (format as date).
}
}
\description{
Reads a table of successive drug delivery records (usually extracted from a pharmacy or a health insurance information system) and creates the data required for the calculation of the polypharmacy indicators by applying various user-defined arguments, incorporating hospital stays into the treatment periods and reconstruct continuous treatment periods by merging quasi continuous and/or overlapping drugs deliveries.
}
\details{
\strong{Variables}:
\itemize{
\item \code{Rx_id}, \code{Cohort_id} and \code{Hosp_id} columns must be of the same class (integer, numeric, character, ...).
\item \code{Rx_drug_deliv}, \code{Hosp_admis} and \code{Hosp_discharge} can be 1) \code{as.Date('yyyy-mm-dd')}, 2) \code{as.character('yyyy-mm-dd')} or 3) \code{as.integer()} where 0 is January 1\ifelse{html}{\out{<sup>st</sup>}}{\eqn{^{st}}}, 1970.
}

\strong{Arguments}:
\itemize{
\item \code{study_start} and \code{study_end} can be 1) \code{as.Date("yyyy-mm-dd")}, 2) \code{as.character("yyyy-mm-dd")} or 3) \code{as.integer()} where 0 is January 1\ifelse{html}{\out{<sup>st</sup>}}{\eqn{^{st}}}, 1970.
}

\strong{Hospital stays}:\cr
Drug availability is assumed to continue during the hospital stay as it is on the day prior admission. The patient is assumed to resume the consumption of the drugs delivered by community pharmacists (as recorded in \code{Rx_deliv}) the day after \code{hosp_discharge}.

\strong{Run-in period}:\cr
A run-in period is necessary to account for the medications that are available to the individuals on the day of \code{study_start}. It is recommended to include a run-in period of about 6 months (e.g. 7 months to account for possible delays) as some drugs are delivered for up to 6 months at once.

\strong{Grace period}:\cr
The grace period is used to determine if two successive deliveries can be considered as a continuous treatment even if there is a gap of several days for which no treatment is apparently available. Two successive deliveries of an identical drug are considered part of a single continuous treatment if the next delivery doesn’t occur more than \code{grace_cst} + (\code{grace_fctr} × \code{Rx_deliv_dur}) days after the end of the latest drug delivery. The availability of extra drugs accumulated over the successive deliveries is accounted for prior to evaluating the duration of the gap between deliveries.

\strong{Performance}\cr
For better performance, date columns are converted to integer numbers.
}
