% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drug_bkdn.R
\name{drug_bkdn}
\alias{drug_bkdn}
\title{Translate combination drug deliveries into several single active ingredients}
\usage{
drug_bkdn(Rx_deliv, Rx_drug_code, Combn_drugs, Combn_drug_code, Combn_act_code)
}
\arguments{
\item{Rx_deliv}{Name of the table listing all prescription drugs deliveries.}

\item{Rx_drug_code}{Column name of \code{Rx_deliv} that contains the combination drug unique identifiers (any format).}

\item{Combn_drugs}{Name of the correspondence table listing all elementary active ingredients that make up each combination drug.}

\item{Combn_drug_code}{Column name of \code{Combn_drugs} that contains the combination drug unique identifiers (same format as \code{Rx_drug_code}).}

\item{Combn_act_code}{Column name of elementary active ingredients that is present in \code{Combn_drugs} (same format as \code{Rx_drug_code}).}
}
\value{
\code{data.table} of the same structure as \code{Rx_deliv}.
}
\description{
Replaces each combination drug into several deliveries of elementary active ingredients according to a user-provided correspondence table.
}
\examples{
Rx_dt <- data.frame(id = c(1, 1, 2, 2, 2),
                    codeDrug = c(159, 753, 123, 456, 789))
SplitCode <- data.frame(code = c(159, 159, 456, 456, 456),
                        split_code = c(1591, 1592, 4567, 4568, 4569))
Rx_split <- drug_bkdn(Rx_deliv = Rx_dt, Rx_drug_code = "codeDrug",
                      Combn_drugs = SplitCode, Combn_drug_code = "code",
                      Combn_act_code = "split_code")
}
