% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drug_bkdn.R
\encoding{UTF-8}
\name{drug_bkdn.verif_args}
\alias{drug_bkdn.verif_args}
\title{Verification}
\usage{
drug_bkdn.verif_args(
  Rx_deliv,
  Rx_drug_code,
  Combn_drugs,
  Combn_drug_code,
  Combn_act_code
)
}
\description{
Arguments verification for \code{\link{drug_bkdn}}.
}
\keyword{internal}
