% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ArgCheck.R
\name{finishArgCheck}
\alias{finishArgCheck}
\title{Argcheck}
\usage{
finishArgCheck(argcheck)
}
\arguments{
\item{argcheck}{Environment created by \code{\link{newArgCheck}}.}
}
\value{
Messages, warnings or errors inserted in \code{\link{newArgCheck}}.
}
\description{
Create an environment where to put errors, messages or warnings.
}
\keyword{internal}
