% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ind_simult.R
\encoding{UTF-8}
\name{ind_simult}
\alias{ind_simult}
\title{Indicator: Simultaneous}
\usage{
ind_simult(
  processed_tab,
  individual_stats = c("mean", "min", "median", "max"),
  stats = c("mean", "sd", "min", "p5", "p10", "p25", "median", "p75", "p90", "p95",
    "max"),
  calendar = FALSE,
  cores = parallel::detectCores(logical = FALSE)
)
}
\arguments{
\item{processed_tab}{Table created by \code{\link{data_process}} function.}

\item{individual_stats}{Statistics to calculate for each drug user. See \emph{Details} for possible values.}

\item{stats}{Statistics to calculate for each \code{individual_stats}. See \emph{Details} for possible values.}

\item{calendar}{\code{TRUE} or \code{FALSE}. Create a table indicating the number of drugs consumed for each day for each user (\code{FALSE} by default).}

\item{cores}{The number of cores to use when executing \code{ind_simult()}. See \code{\link[parallel]{parallel::detectCores}}.}
}
\value{
if \code{calendar} is \code{FALSE}:
\itemize{
\item \code{data.table} indicating each \code{stats} (columns) for each \code{individual_stats} (rows).
}

if \code{calendar} is \code{TRUE}, a list of two (2) elements:
\itemize{
\item \code{indicators}: Table described above.
\item \code{calendar}: Table indicating the number of drugs consumed for each day (only for those who has at least 1 day with 1 drug consumption).
}
}
\description{
Descriptive statistics on daily consumption.
}
\details{
\strong{\code{individual_stats}} & \strong{\code{stats}}: Possible values are
\itemize{
\item \code{'mean'}, \code{'min'}, \code{'median'}, \code{'max'}, \code{'sd'};
\item \code{'pX'} where \emph{X} is a value in ]0, 100];
\item \code{'q1'} = \code{'p25'}, \code{'q2'} = \code{'p50'} = \code{'median'}, \code{q3} = \code{'p75'}.
}
}
\examples{
dt_process <- data_process(
  Rx_deliv = data.frame(
    ID = c(1, 1, 1, 2, 2), Code = c('A', 'B', 'C', 'D', 'E'),
    Date = c('2020-01-01', '2020-01-05', '2020-01-10', '2020-01-15', '2020-01-26'),
    Duration = c(20, 15, 10, 5, 3)
  ), Rx_id = 'ID', Rx_drug_code = 'Code', Rx_drug_deliv = 'Date', Rx_deliv_dur = 'Duration',
  cores = 1
)
dt_simult <- ind_simult(dt_process, cores = 1)
dt_calendar <- ind_simult(dt_process, calendar = TRUE, cores = 1)
}
