% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ind_simult.R
\name{ind_simult}
\alias{ind_simult}
\title{Simultaneous indicator}
\usage{
ind_simult(
  processed_tab,
  individual_stats = c("mean", "min", "median", "max"),
  stats = c("mean", "sd", "min", "p5", "p10", "q1", "median", "q3", "p90", "p95",
    "max"),
  calendar = FALSE
)
}
\arguments{
\item{processed_tab}{Table created by \code{data_process()} function.}

\item{individual_stats}{Statistics to calculate for each drug user. See \emph{Details} for possible values.}

\item{stats}{Statistics to calculate for each \code{individual_stats}. See \emph{Details} for possible values.}

\item{calendar}{\code{TRUE} or \code{FALSE}. Create a table indicating the number of drugs consumed for each day (\code{FALSE} by default).}
}
\value{
if \code{calendar} is \code{FALSE}:
\itemize{
\item data.table\verb{indicating each}stats\verb{(columns) for each}individual_stats\verb{(rows).\\cr\\cr if}calendar\code{is}TRUE`, a list of two (2) elements:
\item indicators: Table described above.
\item calendar: Table indicating the number of drugs consumed for each day.
}
}
\description{
Description
}
\details{
**Attributes from \code{\link{data_process}()} function:\cr
}
