% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ind_stdcontinuous.R
\encoding{UTF-8}
\name{ind_stdcontinuous}
\alias{ind_stdcontinuous}
\title{Indicator: Standard Continuous}
\usage{
ind_stdcontinuous(
  processed_tab,
  pdays,
  stats = c("mean", "sd", "min", "p5", "p10", "p25", "median", "p75", "p90", "p95",
    "max")
)
}
\arguments{
\item{processed_tab}{Table created by \code{\link{data_process}} function.}

\item{pdays}{Number of days to create intervals \verb{[min; min+pdays]} and \verb{[max-pdays; max]} where a drug should be consumed to be counted.}

\item{stats}{Statistics to calculate on the drug consumption. See \emph{Details} for possible values.}
}
\value{
\code{list}:
\itemize{
\item \code{indic}: \code{data.table} indicating each \code{stats} (columns).
\item \code{stats_id}: \code{data.table} indicating the number of drugs use for each individual (all cohort).
}
}
\description{
Descriptive statistics.\cr
A drug is counted if there is a least 1 consumption in the interval \verb{[min; min+pdays]} and another in \verb{[max-pdays; max]}. In other words, a drug consumption is considered continuous if there is a consumption at the beginning and at the end of the period.r
}
\details{
\strong{\code{stats}}: Possible values are
\itemize{
\item \code{'mean'}, \code{'min'}, \code{'median'}, \code{'max'}, \code{'sd'};
\item \code{'pX'} where \emph{X} is a value in ]0, 100];
\item \code{'q1'} = \code{'p25'}, \code{'q2'} = \code{'p50'} = \code{'median'}, \code{q3} = \code{'p75'}.
}
}
