% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ind_stdcumul.R
\encoding{UTF-8}
\name{ind_stdcumul}
\alias{ind_stdcumul}
\title{Indicator: Cumulative (multiple medication)}
\usage{
ind_stdcumul(
  processed_tab,
  nPeriod = 1,
  stats = c("mean", "sd", "min", "p5", "p10", "p25", "median", "p75", "p90", "p95",
    "max")
)
}
\arguments{
\item{processed_tab}{Table created by \code{\link{data_process}} function.}

\item{nPeriod}{Integer value greater or equal to 1 and lesser or equal to the total number of days in the study period. If \code{nPeriod} is greater than 1, the study period is divide in \code{nPeriod} subperiod and the total number of drugs consumption would be the average of the periods.}

\item{stats}{Statistics to calculate on the drug consumption. See \emph{Details} for possible values.}
}
\value{
\code{data.table} indicating each \code{stats} (columns).
}
\description{
Descriptive statistics: Sum of different drugs consumed over a given period time.\cr\cr
\href{https://www.irdes.fr/english/issues-in-health-economics/204-polypharmacy-definitions-measurement-and-stakes-involved.pdf}{Cumulative polypharmacy, also known as multiple medication (Hovstadius et al.,2010a), is defined by the sum of different medications administered over a given period of time (Fincke et al., 2005).}
}
\details{
\strong{\code{stats}}: Possible values are
\itemize{
\item \code{'mean'}, \code{'min'}, \code{'median'}, \code{'max'}, \code{'sd'};
\item \code{'pX'} where \emph{X} is a value in ]0, 100];
\item \code{'q1'} = \code{'p25'}, \code{'q2'} = \code{'p50'} = \code{'median'}, \code{q3} = \code{'p75'}.
}

The attribute \code{nRx} is a \code{data.table} indicating the number of drugs consumption per period and the average of these periods.
}
