% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ind_wcumul.R
\encoding{UTF-8}
\name{ind_wcumul}
\alias{ind_wcumul}
\title{Indicator: Weight Cumulative}
\usage{
ind_wcumul(
  processed_tab,
  stats = c("mean", "sd", "min", "p5", "p10", "p25", "median", "p75", "p90", "p95",
    "max")
)
}
\arguments{
\item{processed_tab}{Table created by \code{\link{data_process}} function.}

\item{stats}{Statistics to calculate on the drug consumption. See \emph{Details} for possible values.}
}
\value{
\code{list}:
\itemize{
\item \code{indic}: \code{data.table} indicating each \code{stats} (columns).
\item \code{stats_id}: \code{data.table} indicating the number of drugs use for each individual (all cohort).
}
}
\description{
Description
}
\details{
\strong{\code{stats}}: Possible values are
\itemize{
\item \code{'mean'}, \code{'min'}, \code{'median'}, \code{'max'}, \code{'sd'};
\item \code{'pX'} where \emph{X} is a value in ]0, 100];
\item \code{'q1'} = \code{'p25'}, \code{'q2'} = \code{'p50'} = \code{'median'}, \code{q3} = \code{'p75'}.
}
}
