% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indicators.R
\encoding{UTF-8}
\name{indicators}
\alias{indicators}
\title{Indicators: All selected}
\usage{
indicators(
  processed_tab,
  stats = c("mean", "sd", "min", "p5", "p10", "p25", "median", "p75", "p90", "p95",
    "max"),
  method = c("ind_simult", "ind_stdcumul", "ind_wcumul", "ind_stdcontinuous",
    "ind_ucontinuous"),
  stdconti_pdays,
  simult_ind_stats = c("mean", "min", "median", "max"),
  simult_calendar = FALSE,
  stdcumul_nPeriod = 1,
  cores = parallel::detectCores(logical = FALSE)
)
}
\arguments{
\item{processed_tab}{Table created by \code{\link{data_process}} function.}

\item{stats}{Statistics to calculate on the drug consumption. See \emph{Details} for possible values.}

\item{method}{Indicator functions name to use.}

\item{stdconti_pdays}{\emph{stdcontinuous} method: Number of days to create intervals \verb{[min; min+pdays]} and \verb{[max-pdays; max]} where a drug should be consumed to be counted.}

\item{simult_ind_stats}{\emph{simult} method: Statistics to calculate for each drug user.}

\item{simult_calendar}{\emph{simul} method: \code{TRUE} or \code{FALSE}. Create a table indicating the number of drugs consumed for each day for each user (\code{FALSE} by default).}

\item{stdcumul_nPeriod}{\emph{std_cumul} method: Integer value greater or equal to 1 and lesser or equal to the total number of days in the study period. If \code{nPeriod} is greater than 1, the study period is divide in \code{nPeriod} subperiod and the total number of drugs consumption would be the average of the periods.}

\item{cores}{The number of cores to use when executing \code{ind_simult()}. See \code{\link[parallel]{parallel::detectCores}}.}
}
\value{
\code{list}
}
\description{
Wrapper function for all \emph{Indicator} functions.
}
\details{
\strong{\code{stats}} & \strong{\code{simult_ind_stats}}: Possible values are
\itemize{
\item \code{'mean'}, \code{'min'}, \code{'median'}, \code{'max'}, \code{'sd'};
\item \code{'pX'} where \emph{X} is a value in ]0, 100];
\item \code{'q1'} = \code{'p25'}, \code{'q2'} = \code{'p50'} = \code{'median'}, \code{q3} = \code{'p75'}.
}
}
