% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-functions.R
\name{rmNA}
\alias{rmNA}
\title{Utils}
\usage{
rmNA(x)
}
\arguments{
\item{x}{Vector.}
}
\value{
\code{x} without \code{NA}s or \code{NaN}s.
}
\description{
Delete \code{NA}s and \code{NaN}s.
}
\examples{
rmNA(c(1:5, NA))
rmNA(c(NaN, NA, 1))
rmNA(c(NaN, NA))
}
\keyword{internal}
