% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-functions.R
\name{sunique}
\alias{sunique}
\title{Utils}
\usage{
sunique(x, decreasing = FALSE, na.last = FALSE)
}
\arguments{
\item{x}{Vector to sort and remove duplicates.}

\item{decreasing}{\code{TRUE} or \code{FALSE}.}

\item{na.last}{\code{NA} removes \code{NA}s, \code{TRUE} show \code{NA}s at the end and \code{FALSE} show \code{NA}s at the beginning.}
}
\description{
Combination of \code{sort()} and \code{unique()}.
}
\keyword{internal}
